Store authentication code as password
The previous way it was stored was just abusing the password field as a storage
for lookup, but if someone tried to log in with the account,
user.check_password() would have raised an exception since it was not in the
correct sha format. Now we're basically setting their password temporarily and
giving it to them in the url. The salt is constant (based on secret key setting)
so that we can still do a direct db lookup.